修正注册后没有显示登陆的Bug 重写了Top 10 页面

+++ up/Application/EvaluatePage/UPEvaluateViewController.m
@@ -32,7 +32,7 @@
 - (void)viewWillAppear:(BOOL)animated {
     [super viewWillAppear:animated];
     [UPNetworkHelper sharedInstance].delegate = self;
-    NSLog(@"fffsssssaaaa-->%d", _positionID);
+    NSLog(@"fffsssssaaaa-->%ld", (long)_positionID);
     [[UPNetworkHelper sharedInstance] postPositionSelectWithID:_positionID];
 }
